chore(datalog): rename tables module
