+++ datalog/src/tables.rs
